requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.5
cloudpathlib==0.16.0
cmake==3.26.4
comm==0.1.4
confection==0.1.4
contexttimer==0.3.3
contourpy==1.1.0
controlnet-aux==0.0.5
cycler==0.11.0
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
diffusers==0.26.2
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.6.1
exceptiongroup==1.1.2
executing==1.2.0
fairscale==0.4.4
fastapi==0.100.0
ffmpy==0.3.1
filelock==3.12.2
flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
gitdb==4.0.10
gitpython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==4.0.0
gradio-client==0.7.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.20.3
identify==2.5.33
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
invisible-watermark==0.2.0
iopath==0.1.10
ipython==8.14.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyterlab-widgets==3.0.8
kaggle==1.6.3
kiwisolver==1.4.4
kornia==0.7.0
langcodes==3.3.0
lazy-loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==16.0.6
lpips==0.1.4
markdown==3.4.4
markdown-it-py==2.2.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.7
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.6.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.0.3
pydantic-core==2.3.0
pydeck==0.8.1b0
pydeprecate==0.3.1
pydub==0.25.1
pygments==2.16.1
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.2
pytorch-lightning==1.4.2
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0
referencing==0.29.1
regex==2023.6.3
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.8.11
rsa==4.9
safetensors==0.3.1
salesforce-lavis==1.0.2
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shellingham==1.5.4
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
streamlit==1.30.0
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.5.0
torchvision==0.15.2
tornado==6.4
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
triton==2.0.0
typer==0.9.0
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.23.0
validators==0.22.0
virtualenv==20.25.0
wandb==0.15.8
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
werkzeug==2.3.6
widgetsnbextension==4.0.8
xformers==0.0.20
yarl==1.9.2
zipp==3.16.2
diffusers